Move confdialog to its own directory
