+++ src/confdialog/fgBgColorClass.py
@@ -7,7 +7,7 @@ from PyQt5.QtWidgets import (
     QDialog,
 )
 
-from .confdialog_Helpers import HotkeySelect, bg_classname
+from .helpers import HotkeySelect, bg_classname
 from .forms import settings_forecolor_bgcolor_class
