+++ src/confdialog/forms/.gitignore
@@ -1,0 +1,1 @@
+/*.py
